refactor(table): move TableSpannedCells to table.dart, make mutation private
Address review feedback on TableSpannedCells:

- Move the class to table.dart so only RenderTable, which shares the
  library, can populate it. The constructor, the mark* methods and the
  hasSpannedCells/isSpanned helpers are now private, leaving TableBorder
  with a read-only view: painting can no longer mutate the flags that
  layout derived.
- Trim implementation details (bit packing, divider skipping) out of the
  API doc and keep them as regular comments in the class body.
